backend/requirements.txt
fastapi
uvicorn
python-dotenv
requestspydantic
sqlalchemy
psycopg2-binary
alembic
